OnlineUser

+++ src/com/lingxi/dataform/UserData.java
@@ -10,6 +10,7 @@ import java.util.List;
 public class UserData extends DataBase {
 
     public static List<User> QueryUserData() {
+        if(!init()) return null;
         String sql = "select * from user";
         try {
             return queryRunner.query(connection, sql, new BeanListHandler<>(User.class));
@@ -19,6 +20,7 @@ public class UserData extends DataBase {
     }
 
     public static int UpdateUserData(User user) {
+        if(!init()) return -1;
         String sql="update user set name=?,age=? where id=?";
         try {
             return queryRunner.update(connection, sql, user.getName(),user.getAge(),user.getId());
@@ -28,6 +30,7 @@ public class UserData extends DataBase {
     }
 
     public static boolean DeleteUserData(User user) {
+        if(!init()) return false;
         String sql="delete from user where id=?";
         try {
             queryRunner.update(connection, sql, user.getId());
